requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.1
urllib3==2.3.0
